Move koji_util.py into utils/
*OSBS-8605

Also change files which import koji_util.py
