+++ atomic_reactor/plugins/exit_sendmail.py
@@ -28,7 +28,7 @@ from atomic_reactor.plugins.exit_store_metadata_in_osv3 import StoreMetadataInOS
 from atomic_reactor.plugins.pre_reactor_config import (get_smtp_session, get_koji_session,
                                                        get_smtp, get_koji, get_openshift,
                                                        get_openshift_session, get_koji_path_info)
-from atomic_reactor.koji_util import get_koji_task_owner
+from atomic_reactor.utils.koji import get_koji_task_owner
 from atomic_reactor.util import get_build_json, OSBSLogs, ImageName, df_parser
 from atomic_reactor.constants import PLUGIN_SENDMAIL_KEY
 from osbs.utils import Labels
