+++ atomic_reactor/plugins/post_koji_upload.py
@@ -17,7 +17,7 @@ from atomic_reactor.plugins.pre_reactor_config import (get_openshift_session,
                                                        get_koji_session)
 from atomic_reactor.constants import PLUGIN_KOJI_UPLOAD_PLUGIN_KEY
 from atomic_reactor.util import (get_build_json, ImageName, is_scratch_build)
-from atomic_reactor.koji_util import get_buildroot, get_output, get_output_metadata
+from atomic_reactor.utils.koji import get_buildroot, get_output, get_output_metadata
 from osbs.exceptions import OsbsException
 
 # An output file and its metadata
